[INJIMOB-891]: rename setup file to index file
