+++ shared/openIdBLE/verifierEventHandler.ts
@@ -1,5 +1,5 @@
 import {VerifierDataEvent} from '../tuvali/types/events';
-import {verifier} from '../tuvali/setup';
+import {verifier} from '../tuvali';
 export function subscribe(callback: (event: VerifierDataEvent) => void) {
   return verifier.handleDataEvents(e => {
     callback(e);
