+++ shared/openIdBLE/walletEventHandler.ts
@@ -1,6 +1,6 @@
 import {WalletDataEvent} from '../tuvali/types/events';
 
-import {wallet} from '../tuvali/setup';
+import {wallet} from '../tuvali';
 
 export function subscribe(callback: (event: WalletDataEvent) => void) {
   return wallet.handleDataEvents(e => {
